+++ src/util/Texture.java
@@ -10,12 +10,12 @@ public class Texture {
     private final TextureAtlas mAtlas;
     private final float[] mCoordinates = new float[4];
 
-    public Texture(TextureAtlas pAtlas, float pX, float pY, float pWidth, float pHeight) {
+    public Texture(TextureAtlas pAtlas, double pX, double pY, double pWidth, double pHeight) {
         mAtlas = Objects.requireNonNull(pAtlas);
-        mCoordinates[0] = pX;
-        mCoordinates[1] = pY;
-        mCoordinates[2] = pWidth;
-        mCoordinates[3] = pHeight;
+        mCoordinates[0] = (float)pX;
+        mCoordinates[1] = (float)pY;
+        mCoordinates[2] = (float)pWidth;
+        mCoordinates[3] = (float)pHeight;
     }
 
     void load(int pUnit) {
